Revising context timeout logic when requesting a new token (#134)
The context we had previously attached to our token refresh call had a
timeout of 10 seconds. This means that our token would only be valid for
that long. We are adjusting our context to only wrap the requests that
retrieve a new token with a timeout to still failover if the request
does not complete, but not to nullify the token itself.


Also adding consistency with how we are reading and writing k8s secrets.