Fix Cilium network flows being incomplete (#138)
Noticed an issue where some flows from Hubble Relay were incomplete or
not filled in with what we require in order to make sense of a flow log.
This change addresses that by saying if any of what we need is nil we
should consider this flow not worth attempting to resolve.

The values we are checking are either required for flow resolution or
will never be null based on their assignment (DestinationService).
Notice we are not checking for


// The CiliumNetworkPolicies allowing the egress of the flow.
EgressAllowedBy 
// The CiliumNetworkPolicies allowing the ingress of the flow.
IngressAllowedBy []*Policy 
// The CiliumNetworkPolicies denying the egress of the flow.
EgressDeniedBy []*Policy 
// The CiliumNetworkPolicies denying the ingress of the flow.
IngressDeniedBy []*Policy 


The reason being is because flows are not guaranteed to have any of this
if someone doesn't actually have any network policies configured in the
first place, so even though they are "required" as part of our
ciliumFlow proto message, we will not block pushing a flow downstream if
these are nil or empty.