Fix gRPC logger to log structured fields (#141)
- Reimplement gRPC logger as a Zap Core wrapper instead of a Core hook,
so it can log the structured log fields.
- Replace `SugaredLogger` with `Logger` in the whole codebase to improve
efficiency and eliminate formatting bugs.
- Encode each entry into a string immediately when logged, before
buffering, and never refer to the input Entry and Fields afterwards, as
it is reduces the buffer's memory usage and it allows the entry and
fields to be garbage-collected while a log entry is buffered.
- Fix log buffering to stop buffering when reaching the buffer capacity.
- Eliminate a channel leak by not replacing the `done` channel without
closing it when updating the gRPC connection and stream client.