Simplify the sleep timer/reset timer in manageStreams & fix ConnectStreams channel leak (#164)
There are 2 clear cut-points: the exponential backoff (sleepTimer) and
the reset backoff (resetTimer). These can both be modeled as exponential
backoffs. Here's the behavior that we get when using this double
backoff. The blue line is how much we are sleeping. The light pink line
is how many times the resetCount (severe error) has triggered. As you
can see, it resets the backoff to baseline.


![image](https://github.com/user-attachments/assets/2a4c634d-3dbe-4b47-be80-c1891ad4bf76)