Cancel gRPC stream contexts when keepalives fail (#168)
We found that the resource stream still did not recover when the HTTP/2
connection was closed, despite adding application-level keepalives,
because a keepalive failure was not canceling the stream's context,
which would have canceled the k8s watcher. For streams other than
resources, their contexts were canceled in other places so they did not
have the same issue.