Added "LongSuccessBackoffReset" path in backoff function
Let's say we fail 7 times, then succeed for 45 minutes minutes before
failing again. Sometimes this 45 minute period of good behavior should
be considered a success, even though the function didn't return "no
error".

ActionTimeToConsiderSuccess is used in the case that the happy-path of our
function is blocking forever. A function like this may have issues that
require exponentialBackoff but should be considered recovered after
the action runs for ActionTimeToConsiderSuccess.

If you don't configure an ActionTimeToConsiderSuccess then this feature
is ignored